refactor(db): dedupe getTsconfigAliases
