+++ cli/utils/db.mjs
@@ -1,4 +1,5 @@
-import { join } from 'pathe'
+import { readFile } from 'node:fs/promises'
+import { join, resolve } from 'pathe'
 import { createStorage } from 'unstorage'
 import fsDriver from 'unstorage/drivers/fs'
 
@@ -31,4 +32,19 @@ export async function getNextMigrationNumber() {
     .pop() ?? 0
 
   return (lastSequentialMigrationNumber + 1).toString().padStart(4, '0')
+}
+
+export async function getTsconfigAliases(cwd) {
+  try {
+    const tsconfig = JSON.parse(await readFile(join(cwd, '.nuxt/tsconfig.json'), 'utf-8'))
+    const paths = tsconfig.compilerOptions?.paths || {}
+    const alias = {}
+    for (const [key, values] of Object.entries(paths)) {
+      const resolvedPath = key.endsWith('/*') ? values[0].replace(/\/\*$/, '') : values[0]
+      alias[key.replace(/\/\*$/, '')] = resolve(join(cwd, '.nuxt'), resolvedPath)
+    }
+    return alias
+  } catch {
+    return {}
+  }
 }
